requirements.txt
Django==3.0.5
djangorestframework==3.11.0
pdfkit==0.6.1
androguard==3.3.5
lxml==4.5.0
rsa==4.0
biplist==1.0.3
requests==2.23.0
bs4==0.0.1
colorlog==4.1.0
macholib==1.14
google-play-scraper==0.0.2.2
whitenoise==5.0.1d
waitress==1.4.3
gunicorn==20.0.4
frida==12.8.20
psutil==5.7.0
shelljob==0.5.6
asn1crypto==1.3.0
oscrypto==1.1.0
distro==1.5.0
psycopg2-binary==2.8.5
IP2Location==8.4.1
# Do not update the following
apkid==2.1.0
cryptography<2.5,>=2.1.4 # pyup: ignore
pyOpenSSL==19.0.0 # pyup: ignore
http-tools==1.1.0
django-allauth==0.41.0
django-guardian==2.2.0
